inital commit for nvim configuration rework
