--- .config/nvim/dot_netrwhist
@@ -1,4 +1,0 @@
-let g:netrw_dirhistmax  =10
-let g:netrw_dirhistcnt =2
-let g:netrw_dirhist_2='/home/cole/.config/nvim'
-let g:netrw_dirhist_1='/home/cole/Projects/SPIKE-AzureContainerApps'
