+++ .config/nvim/lua/cocodust/base/lsp/utils.lua
@@ -1,0 +1,18 @@
+local M = {}
+
+function M.capabilites()
+  local capabilities = vim.lsp.protocol.make_client_capabilities()
+  return require("cmp_nvim_lsp").default_capabilities(capabilities)
+end
+
+function M.on_attach(on_attach)
+  vim.api.nvim_create_autocmd("LspAttach", {
+    callback = function(args)
+      local bufnr = args.buf
+      local client = vim.lsp.get_client_by_id(args.data.client_id)
+      on_attach(client, bufnr)
+    end
+  })
+end
+
+return M
