+++ .config/nvim/lua/cocodust/base/mason.lua
@@ -1,0 +1,35 @@
+return {
+  "williamboman/mason.nvim",
+  build = ":MastonUpdate",
+  cmd = "Mason",
+  keys = {
+    {
+      "<leader>cm",
+      "<cmd>Mason<CR>",
+      desc = "[C]heck [M]ason",
+    },
+  },
+  opts = {
+    ensure_installed = {
+      "gitlint",
+    },
+  },
+  config = function(_, opts)
+    require("mason").setup(opts)
+    local mr = require("mason-registry")
+    local function ensure_installed()
+      for _, tool in ipairs(opts.ensure_installed) do
+        local p = mr.get_package(tool)
+        if not p:is_installed() then
+          p:install()
+        end
+      end
+    end
+
+    if mr.refresh then
+      mr.refresh(ensure_installed)
+    else
+      ensure_installed()
+    end
+  end
+}
