+++ .config/nvim/lua/cocodust/base/snippets.lua
@@ -1,0 +1,13 @@
+return {
+  "L3MON4D3/LuaSnip",
+  dependencies = {
+    "rafamadriz/friendly-snippets",
+    config = function()
+      require("luasnip.loaders.from_vscode").lazy_load()
+    end,
+    opts = {
+      history = true,
+      delete_check_events = "TextChanged",
+    },
+  }
+}
