+++ .config/nvim/lua/cocodust/config/lazy-modules.lua
@@ -1,0 +1,6 @@
+return {
+	{ import = "cocodust.base" },
+	{ import = "cocodust.plugins" },
+	{ import = "cocodust.plugins.random" },
+	{ import = "cocodust.plugins.language" },
+}
