--- .config/nvim/lua/cocodust/plugins/autopair.lua
@@ -1,32 +1,0 @@
-return {
-	{
-		"windwp/nvim-autopairs",
-		lazy = false,
-		enabled = true,
-		opts = {
-			check_ts = true,
-		},
-	},
-	{
-		"windwp/nvim-ts-autotag",
-		ft = {
-			"html",
-			"javascript",
-			"typescript",
-			"javascriptreact",
-			"typescriptreact",
-			"svelte",
-			"vue",
-			"tsx",
-			"jsx",
-			"rescript",
-			"xml",
-			"php",
-			"markdown",
-			"astro",
-			"glimmer",
-			"handlebars",
-			"hbs",
-		},
-	},
-}
