--- .config/nvim/lua/cocodust/plugins/commentary.lua
@@ -1,7 +1,0 @@
-return {
-	{
-		"tpope/vim-commentary",
-		lazy = false,
-		priority = 1001,
-	},
-}
