--- .config/nvim/lua/cocodust/plugins/gitgutter.lua
@@ -1,12 +1,0 @@
-local icons = require("cocodust.icons")
-return {
-	{
-		"airblade/vim-gitgutter",
-		lazy = false,
-		init = function()
-			vim.g.gitgutter_sign_added = icons.Git.Added.Icon.Box_Plus
-			vim.g.gitgutter_sign_modified = icons.Git.Modified.Icon.Pencil
-			vim.g.gitgutter_sign_removed = icons.Git.Deleted.Icon.Minus
-		end,
-	},
-}
