+++ .config/nvim/lua/cocodust/plugins/language/comments.lua
@@ -1,0 +1,8 @@
+return {
+  {
+    "JoosepAlviste/nvim-ts-context-commentstring",
+    opts = {
+      enable_autocmd = false,
+    },
+  },
+}
