+++ .config/nvim/lua/cocodust/plugins/language/css.lua
@@ -1,0 +1,18 @@
+return {
+  {
+    "nvim-treesitter/nvim-treesitter",
+    opts = function(_, opts)
+      vim.list_extend(opts.ensure_installed, { "css" })
+    end,
+  },
+  {
+    "neovim/nvim-lspconfig",
+    opts = {
+      servers = {
+        cssls = {},
+        cssmodules_ls = {},
+        tailwindcss = {},
+      },
+    },
+  },
+}
