+++ .config/nvim/lua/cocodust/plugins/language/frontend/init.lua
@@ -1,0 +1,4 @@
+return {
+  "andymass/vim-matchup",
+  lazy = false
+}
