+++ .config/nvim/lua/cocodust/plugins/language/lua.lua
@@ -1,0 +1,31 @@
+return {
+  {
+    "nvim-treesitter/nvim-treesitter",
+    opts = function(_, opts)
+      vim.list_extend(opts.ensure_installed, { "lua", "luadoc", "luap" })
+    end,
+  },
+  {
+    "williamboman/mason.nvim",
+    opts = function(_, opts)
+      vim.list_extend(opts.ensure_installed, { "stylua" })
+    end,
+  },
+  {
+    "neovim/nvim-lspconfig",
+    opts = {
+      servers = {
+        lua_ls = {
+          settings = {
+            Lua = {
+              diagnostics = {
+                globals = { "vim" }
+              }
+            }
+          }
+        },
+      },
+      setup = {}
+    }
+  }
+}
