+++ .config/nvim/lua/cocodust/plugins/language/powershell.lua
@@ -1,0 +1,8 @@
+return {
+  "neovim/nvim-lspconfig",
+  opts = {
+    servers = {
+      powershell_es = {},
+    },
+  },
+}
