--- .config/nvim/lua/cocodust/plugins/lsp/formatting.lua
@@ -1,47 +1,0 @@
--- Remove this dependency
-local Util = require("lazy.core.util")
-
-local M = {}
-
-M.autoformat = true
-
-function M.toggle()
-	M.autoformat = not M.autoformat
-	if M.autoformat then
-		Util.info("Enabled format on save", { title = "Format" })
-	else
-		Util.warn("Disabled format on save", { title = "Format" })
-	end
-end
-
-function M.format()
-	local bufnr = vim.api.nvim_get_current_buf()
-	local ft = vim.bo[bufnr].filetype
-	local have_nls = #require("null-ls.sources").get_available(ft, "NULL_LS_FORMATTING") > 0
-
-	vim.lsp.buf.format({
-		bufnr = bufnr,
-		filter = function(client)
-			if have_nls then
-				return client.name == "null-ls"
-			end
-			return client.name ~= "null-ls"
-		end,
-	})
-end
-
-function M.on_attach(client, bufnr)
-	if client.server_capabilities.documentFormattingProvider then
-		vim.api.nvim_create_autocmd("BufWritePre", {
-			group = vim.api.nvim_create_augroup("LspFormat" .. bufnr, {}),
-			buffer = bufnr,
-			callback = function()
-				if M.autoformat then
-					M.format()
-				end
-			end,
-		})
-	end
-end
-
-return M
