--- .config/nvim/lua/cocodust/plugins/matchup.lua
@@ -1,5 +1,0 @@
-return {
-	-- match html tags
-	"andymass/vim-matchup",
-	lazy = false,
-}
