+++ .config/nvim/lua/cocodust/plugins/random/colorizer-css.lua
@@ -1,0 +1,14 @@
+return {
+  "NvChad/nvim-colorizer.lua",
+  config = function()
+    require("colorizer").setup({
+      user_default_options = {
+        tailwind = true
+      },
+      filetypes = {
+        "css",
+        "vue",
+      },
+    })
+  end,
+}
