+++ .config/nvim/lua/cocodust/plugins/random/matchup.lua
@@ -1,0 +1,4 @@
+return {
+  "anymass/vim-matchup",
+  lazy= false,
+}
