--- .config/nvim/lua/cocodust/plugins/util.lua
@@ -1,3 +1,0 @@
-return {
-	{ "nvim-lua/plenary.nvim" },
-}
